Rebuild as a static site
I built this with metalsmith as a static site to be hosted with GitHub
Pages. Using metalsmith let me write the code in CoffeeScript, and the
frontend code in Pug, Stylus and CoffeeScript.

Metalsmith also lets me write content in Markdown, and is extensible
enough to let me expand to add math symbols and possibly Fountain later.
It also minifies everything and handles most of the stuff I don't want
to do manually every time, which is great.

Seems cool!