Add SVGO and Fountain
SVGO is an SVG optimiser that helps keep the logo efficient. And it's
pretty efficient, at only 360-something bytes.

Fountain is a screenplay language like Markdown, which I'm using to
render some of my work. I just wrote and published metalsmith-fountain,
which I've included in the project (yay!). I've also added a copy of
John August's Big Fish for testing.